docs(model): fully document `channel::message`
Denies missing docs on it and any of it's submodules
