+++ model/src/channel/message/reaction.rs
@@ -1,10 +1,14 @@
 use crate::channel::ReactionType;
 use serde::{Deserialize, Serialize};
 
+/// Message reaction struct.
 #[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, Serialize)]
 pub struct MessageReaction {
+    /// Ammount of reactions this emoji has.
     pub count: u64,
+    /// Emoji of this reaction.
     pub emoji: ReactionType,
+    /// Whether we have reacted with this emoji.
     pub me: bool,
 }
